[svn-r30292] Description:     Correct failure on older Windows compilers.
Tested on:
    MacOSX/64 10.11.5 (amazon) w/serial & parallel
    (h5committest forthcoming)

+++ test/error_test.c
@@ -284,7 +284,7 @@ error_stack(void)
     /* Make it push error, force this function to fail */
     if((err_num = H5Eget_num(ERR_STACK)) == 0) {
         H5Epush(ERR_STACK, __FILE__, FUNC_error_stack, __LINE__, ERR_CLS, ERR_MAJ_API, ERR_MIN_GETNUM,
-                "Get number test failed, returned %zd", err_num);
+                "Get number test failed, returned %d", (int)err_num);
         goto error;
     } /* end if */
